+++ Source/utils/PathProducer.h
@@ -59,7 +59,7 @@ struct PathProducer : juce::Thread
                 auto success = fftDataGenerator.getFFTData (std::move (fftData));
                 jassert (success);
                 updateRenderData (renderData, fftData, fftSize / 2, static_cast<float> (LOOP_DELAY) * decayRateInDbPerSec.load() / 1000.f);
-                pathGenerator.generatePath (renderData, fftBounds, fftSize, getBinWidth(), negativeInfinity.load(), maxDecibels.load());
+                pathGenerator.generatePath (renderData, fftBounds, fftSize, static_cast<float>(getBinWidth()), negativeInfinity.load(), maxDecibels.load());
             }
             wait (LOOP_DELAY);
         }
@@ -71,7 +71,7 @@ struct PathProducer : juce::Thread
         fftDataGenerator.changeOrder (o);
         renderData.clear();
         auto fftSize = getFFTSize();
-        renderData.resize (fftSize / 2 + 1, negativeInfinity.load());
+        renderData.resize (static_cast<size_t>(fftSize / 2 + 1), negativeInfinity.load());
 
         bufferForGenerator.setSize (1, fftSize, false, false, true);
         bufferForGenerator.clear();
